Adds more detailed documentation to storybook, refactors sample step data for dryer code.
