+++ packages/design-system/src/components/web-components/ds-step-list/ds-step-list.tsx
@@ -3,20 +3,8 @@ import { StepListProps } from '../../StepList/StepList';
 import { StepList } from '../../StepList';
 import { parseBooleanAttr } from '../wrapperUtils';
 
-/*
-Todos:
-  1. Verify the handling of the `component` attribute.
-  2. Verify the handling of `onStepLinkClick`.
-  3. Ensure full test coverage in `ds-step-list.tests.tsx`.
-*/
-
 const attributes = [
   'steps',
-  /*
-    The component prop in our React component is used to render a specific React component 
-    for all link elements when provided.
-    */
-  // 'component', // StepLinkProps['component']
   'show-sub-sub-steps',
   'completed-text',
   'edit-text',
@@ -41,24 +29,19 @@ const Wrapper = ({
   actionsLabelText,
   substepsLabelText,
   ...otherProps
-}: WrapperProps) => {
-  // console.log('steps', steps)
-  const parsedShowSubSubSteps = parseBooleanAttr(showSubSubSteps);
-
-  return (
-    <StepList
-      steps={typeof steps === 'string' ? JSON.parse(steps) : steps}
-      showSubSubSteps={parsedShowSubSubSteps}
-      completedText={completedText}
-      editText={editText}
-      resumeText={resumeText}
-      startText={startText}
-      actionsLabelText={actionsLabelText}
-      substepsLabelText={substepsLabelText}
-      {...otherProps}
-    />
-  );
-};
+}: WrapperProps) => (
+  <StepList
+    steps={typeof steps === 'string' ? JSON.parse(steps) : steps}
+    showSubSubSteps={parseBooleanAttr(showSubSubSteps)}
+    completedText={completedText}
+    editText={editText}
+    resumeText={resumeText}
+    startText={startText}
+    actionsLabelText={actionsLabelText}
+    substepsLabelText={substepsLabelText}
+    {...otherProps}
+  />
+);
 
 /* eslint-disable @typescript-eslint/no-namespace */
 declare global {
